fix: rename DeleteQuotaForTierCommand to DeleteTierQuotaDefinitionCommand
