--- Modules/Quotas/src/Quotas.Application/Tiers/Commands/DeleteQuotaForTier/DeleteQuotaForTierCommand.cs
@@ -1,14 +1,0 @@
-﻿using MediatR;
-
-namespace Backbone.Modules.Quotas.Application.Tiers.Commands.DeleteQuotaForTier;
-public class DeleteQuotaForTierCommand : IRequest
-{
-    public DeleteQuotaForTierCommand(string tierId, string tierQuotaDefinitionId)
-    {
-        TierId = tierId;
-        TierQuotaDefinitionId = tierQuotaDefinitionId;
-    }
-
-    public string TierId { get; set; }
-    public string TierQuotaDefinitionId { get; set; }
-}
