+++ Modules/Quotas/src/Quotas.Application/Tiers/Commands/DeleteTierQuotaDefinition/DeleteTierQuotaDefinitionCommandValidator.cs
@@ -1,0 +1,12 @@
+﻿using Enmeshed.BuildingBlocks.Application.FluentValidation;
+using FluentValidation;
+
+namespace Backbone.Modules.Quotas.Application.Tiers.Commands.DeleteTierQuotaDefinition;
+public class DeleteTierQuotaDefinitionCommandValidator : AbstractValidator<DeleteTierQuotaDefinitionCommand>
+{
+    public DeleteTierQuotaDefinitionCommandValidator()
+    {
+        RuleFor(c => c.TierId).DetailedNotEmpty();
+        RuleFor(c => c.TierQuotaDefinitionId).DetailedNotEmpty();
+    }
+}
