+++ Modules/Quotas/test/Quotas.Application.Tests/Tests/Quotas/DeleteTierQuotaDefinition/HandlerTests.cs
@@ -1,6 +1,6 @@
 ﻿using Backbone.Modules.Quotas.Application.Infrastructure.Persistence.Repository;
 using Backbone.Modules.Quotas.Application.IntegrationEvents.Outgoing;
-using Backbone.Modules.Quotas.Application.Tiers.Commands.DeleteQuotaForTier;
+using Backbone.Modules.Quotas.Application.Tiers.Commands.DeleteTierQuotaDefinition;
 using Backbone.Modules.Quotas.Domain.Aggregates.Identities;
 using Backbone.Modules.Quotas.Domain.Aggregates.Metrics;
 using Backbone.Modules.Quotas.Domain.Aggregates.Tiers;
@@ -11,7 +11,7 @@ using FluentAssertions.Execution;
 using Microsoft.Extensions.Logging;
 using Xunit;
 
-namespace Backbone.Modules.Quotas.Application.Tests.Tests.Quotas.DeleteQuotaForTier;
+namespace Backbone.Modules.Quotas.Application.Tests.Tests.Quotas.DeleteTierQuotaDefinition;
 public class HandlerTests
 {
     private readonly IEventBus _eventBus;
@@ -35,7 +35,7 @@ public class HandlerTests
         var period = QuotaPeriod.Month;
         tier.CreateQuota(metricKey, max, period);
 
-        var command = new DeleteQuotaForTierCommand(tier.Id, tier.Quotas.ElementAt(0).Id);
+        var command = new DeleteTierQuotaDefinitionCommand(tier.Id, tier.Quotas.ElementAt(0).Id);
 
         var tierRepository = A.Fake<ITiersRepository>();
         A.CallTo(() => tierRepository.Find(tierId, A<CancellationToken>._, A<bool>._)).Returns(tier);
@@ -68,7 +68,7 @@ public class HandlerTests
         tier.CreateQuota(metricKey, max, period);
         tier.CreateQuota(metricKey, max, period);
 
-        var command = new DeleteQuotaForTierCommand(tier.Id, tier.Quotas.ElementAt(0).Id);
+        var command = new DeleteTierQuotaDefinitionCommand(tier.Id, tier.Quotas.ElementAt(0).Id);
 
         var tierRepository = A.Fake<ITiersRepository>();
         A.CallTo(() => tierRepository.Find(tierId, A<CancellationToken>._, A<bool>._)).Returns(tier);
@@ -96,7 +96,7 @@ public class HandlerTests
 
         tier.CreateQuota(MetricKey.NumberOfSentMessages, 5, QuotaPeriod.Month);
 
-        var command = new DeleteQuotaForTierCommand(tier.Id, tier.Quotas.ElementAt(0).Id);
+        var command = new DeleteTierQuotaDefinitionCommand(tier.Id, tier.Quotas.ElementAt(0).Id);
 
         var tierRepository = A.Fake<ITiersRepository>();
         A.CallTo(() => tierRepository.Find(tierId, A<CancellationToken>._, A<bool>._)).Returns(tier);
